vultr-backend/requirements.txt
# FastAPI Application Requirements - EXACT WORKING VERSIONS
# Generated from working environment on 2025-07-08

# Core FastAPI dependencies
fastapi==0.115.14
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Pydantic for data validation
pydantic==2.11.7
pydantic-core==2.33.2

# HTTP client for external API calls
requests==2.32.4
httpx==0.28.1

# PDF processing
pdfplumber==0.11.7
pdfminer.six==20250506

# AI and ML dependencies - EXACT WORKING VERSIONS
crewai==0.140.0
groq==0.29.0
gitingest==0.1.5

# CrewAI core dependencies
openai==1.93.0
instructor==1.9.0
litellm==1.72.6

# Environment variables
python-dotenv==1.1.1

# Date and time handling
python-dateutil==2.9.0.post0

# JSON processing
json-repair==0.25.2

# Async support
anyio==4.9.0
sniffio==1.3.1

# Core Python dependencies
typing-extensions==4.14.1
annotated-types==0.7.0

# Security and cryptography
cryptography==45.0.5

# File handling and utilities
pathspec==0.12.1
packaging==25.0

# Network and HTTP
aiohttp==3.12.13
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0

# Development and testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1 